backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.6.0
gunicorn==21.2.0
waitress==2.1.2